move Header and Login to components folder
