+++ src/features/Authentication/components/Header/Header.styled.tsx
